+++ src/modules/balance/ModuleBalanceCard.vue
@@ -1,11 +1,18 @@
 <template>
   <div class="component--wallet-card">
-    <div class="background-image">
-      <img src="@/assets/images/snippets/wallet-card.png" />
+    <div class="mew-card">
+      <img
+        :src="'https://mewcard.mewapi.io/?address=' + address"
+        alt="MEW Card"
+        @load="
+          animateMewCard();
+          animateBlockie();
+        "
+      />
     </div>
     <div class="info-container px-4 pt-5 pb-3">
       <div class="d-flex">
-        <div class="blockie-img">
+        <div ref="blockie" class="blockie-img">
           <mew-blockie
             :address="address"
             :size="8"
@@ -20,16 +27,23 @@
         </div>
         <div class="ml-4">
           <div class="font-weight-medium d-flex align-center">
-            <div>MY ACCOUNT VALUE</div>
+            <div class="text-shadow font-weight-bold">MY ACCOUNT VALUE</div>
           </div>
-          <div class="headline font-weight-bold monospace">
+          <div
+            v-show="convertedBalance !== 'undefinedNaN'"
+            class="text-shadow headline font-weight-bold monospace"
+          >
             {{ convertedBalance }}
           </div>
         </div>
       </div>
       <div class="component--address d-flex align-center mt-1">
-        <div class="monospace full-address">{{ getChecksumAddressString }}</div>
-        <div class="monospace last-four">{{ lastFour }}</div>
+        <div class="text-shadow monospace full-address font-weight-bold">
+          {{ address }}
+        </div>
+        <div class="text-shadow monospace last-four font-weight-bold">
+          {{ lastFour }}
+        </div>
       </div>
       <!-- <div class="mb-2">OWNED 3 DOMAINS ></div> -->
       <div class="d-flex align-center mt-2">
@@ -39,6 +53,7 @@
               src="@/assets/images/icons/icon-print-light.png"
               alt="Print Wallet"
               height="20"
+              class="drop-shadow"
             />
           </v-btn>
           <v-btn icon @click="copyAddress">
@@ -46,6 +61,7 @@
               src="@/assets/images/icons/icon-copy-light.png"
               alt="Print Wallet"
               height="20"
+              class="drop-shadow"
             />
           </v-btn>
         </div>
@@ -73,6 +89,7 @@
 </template>
 
 <script>
+import anime from 'animejs/lib/anime.es.js';
 import BalanceAddressSwitch from './components/BalanceAddressSwitch';
 import BalanceAddressPaperWallet from './components/BalanceAddressPaperWallet';
 import BalanceAddressQrCode from './components/BalanceAddressQrCode';
@@ -123,7 +140,33 @@ export default {
       }`;
     }
   },
+  mounted() {
+    //this.animateBlockie();
+  },
   methods: {
+    animateBlockie() {
+      const el = document.querySelector('.blockie-img');
+      el.style.transform = 'scale(0)';
+      anime({
+        targets: el,
+        keyframes: [{ scale: 0 }, { scale: 3 }, { scale: 1 }],
+        delay: 2000,
+        duration: 2000
+      });
+    },
+    animateMewCard() {
+      const el = document.querySelector('.mew-card');
+      //el.style.transform = 'scale(6)';
+      el.style.opacity = 0;
+      anime({
+        targets: el,
+        //scale: 1,
+        opacity: 1,
+        delay: 1800,
+        duration: 500,
+        easing: 'easeInOutQuad'
+      });
+    },
     closeChangeAddress() {
       this.openChangeAddress = false;
     },
@@ -140,10 +183,15 @@ export default {
 
 <style lang="scss" scoped>
 .component--wallet-card {
+  background-color: rgba(255, 255, 255, 0.1);
+  border-radius: 10px;
   position: relative;
   width: 100%;
 }
-.background-image {
+.mew-card {
+  opacity: 0;
+  border-radius: 10px;
+  overflow: hidden;
   position: absolute;
   top: 0;
   left: 0;
@@ -196,6 +244,14 @@ export default {
 
 .last-four {
   margin-left: -4px;
+}
+
+.text-shadow {
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 1);
+}
+
+.drop-shadow {
+  filter: drop-shadow(1px 1px 2px rgba(0, 0, 0, 0.8));
 }
 </style>